Add static busybox back
Used on a lot of samsung devices.
We've no recovery partition and cannot make sure that all users using a specific base
for installing cyanogenmod the first time.
We're running a special updater shellscript which is doing a lot of shit using busybox.
We prefer to have a static busybox in this case.

Change-Id: I491919b869c0469ccb092b8085c7dab53dc72f20